pyproject.toml
[tool.poetry]
name = "python-ai"
version = "0.1.0"
description = ""
authors = ["Aleksandr Liadov <aleksandr.liadov@stellantis.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
github-action-utils = "^1.1.0"
tenacity = "^8.3.0"
pydantic = "^2.7.2"
pyyaml = "^6.0.1"
openai = "^1.30.5"
pygithub = "^2.3.0"
tiktoken = "^0.7.0"
python-dotenv = "^1.0.1"
python-box = "^7.1.1"
mistralai = "^0.4.1"
text-generation = "^0.7.0"
httpx = "^0.27.0"
sentencepiece = "^0.2.0"
protobuf = "^5.27.2"
huggingface-hub = "^0.23.4"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
